requirements.txt
aiogram==3.3.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.120.0
python-dotenv==1.0.1
aiohttp==3.9.3
certifi==2024.2.2
urllib3==2.2.1 